feat: removed v2 and updated v1
