+++ nodes/JsonValidator/JsonValidator.node.ts
@@ -1,5 +1,4 @@
 import { INodeTypeBaseDescription, IVersionedNodeType, VersionedNodeType } from 'n8n-workflow';
-import { JsonValidatorV2 } from './v2/JsonValidatorV2.node';
 import { JsonValidatorV1 } from './v1/JsonValidatorV1.node';
 
 export class JsonValidator extends VersionedNodeType {
@@ -11,12 +10,11 @@ export class JsonValidator extends VersionedNodeType {
 			icon: 'file:ajv.svg',
 			description: 'Validate data using a JSON Schema',
 			subtitle: 'Validate JSON using AJV',
-			defaultVersion: 2,
+			defaultVersion: 1,
 		};
 
 		const nodeVersions: IVersionedNodeType['nodeVersions'] = {
 			1: new JsonValidatorV1(baseDescription),
-			2: new JsonValidatorV2(baseDescription),
 		};
 
 		super(nodeVersions, baseDescription);
