+++ nodes/JsonValidator/v1/JsonValidatorV1.node.ts
@@ -4,11 +4,11 @@ import {
 	INodeType,
 	INodeTypeBaseDescription,
 	INodeTypeDescription,
-	NodeOperationError,
 } from 'n8n-workflow';
 
 import Ajv, { type SchemaObject } from 'ajv';
 import addFormats from 'ajv-formats';
+import addErrors from 'ajv-errors';
 
 export class JsonValidatorV1 implements INodeType {
 	description: INodeTypeDescription;
@@ -20,23 +20,41 @@ export class JsonValidatorV1 implements INodeType {
 			defaults: {
 				name: 'Json Validator',
 			},
-			description: 'Validate data using a JSON Schema',
 			inputs: ['main'],
 			outputs: ['main'],
 			properties: [
 				{
-					displayName: 'Scheme (JSON)',
-					name: 'scheme',
-					type: 'json',
-					default: {},
-					description: 'A valid ajv Scheme',
+					displayName: `Validate data using a JSON Schema.<br />
+						A valid draft-07 schema. Visit <a href='https://JSON-schema.org/draft-07/' target='_blank'>JSON-schema.org</a> to learn how to describe your validation rules in JSON Schemas.<br />
+						It does make use of the errors properties of ajv-errors, learn more about it <a href='https://github.com/ajv-validator/ajv-errors'>here</a>`,
+					name: 'notice',
+					type: 'notice',
+					default: '',
 				},
 				{
-					displayName: 'Input Field',
-					name: 'inputField',
-					type: 'string',
-					default: 'json',
-					description: 'The name of the input field containing the JSON data to validate',
+					displayName: 'JSON Schema',
+					name: 'schema',
+					type: 'json',
+					required: true,
+					typeOptions: {
+						alwaysOpenEditWindow: true,
+					},
+					default: JSON.stringify(
+						{
+							type: 'object',
+							properties: {
+								foo: { type: 'integer' },
+								bar: { type: 'string' },
+							},
+							required: ['foo'],
+							additionalProperties: false,
+						},
+						undefined,
+						2,
+					),
+					placeholder: '',
+					description:
+						'A valid draft-07 schema. Visit https://JSON-schema.org/draft-07 to learn how to describe your validation rules in JSON Schemas.',
 				},
 			],
 		};
@@ -46,19 +64,38 @@ export class JsonValidatorV1 implements INodeType {
 		/**
 		 * Initiate AVJ
 		 */
-		const AJV = new Ajv();
-		addFormats(AJV);
+		const ajv = new Ajv({ allErrors: true });
+		addFormats(ajv);
+		addErrors(ajv);
 
 		// Get the JSON schema defined by the user
-		const scheme = this.getNodeParameter('scheme', 0, undefined, {
+		const schema = this.getNodeParameter('schema', 0, undefined, {
 			ensureType: 'json',
 		}) as SchemaObject;
 
-		// Get the input field specified by the user
-		const inputField = this.getNodeParameter('inputField', 0) as string;
+		try {
+			ajv.validateSchema(schema);
+			if (ajv.errors) {
+				return this.prepareOutputData([
+					{
+						json: {
+							error: `Invalid JSON Schema: ${ajv.errorsText(ajv.errors)}`,
+						},
+					},
+				]);
+			}
+		} catch (error) {
+			return this.prepareOutputData([
+				{
+					json: {
+						error: `Invalid JSON Schema ${error}`,
+					},
+				},
+			]);
+		}
 
-		//Compile scheme
-		const validate = AJV.compile(scheme);
+		// Compile schema
+		const validate = ajv.compile(schema);
 
 		// Get the input data for validation
 		const items = this.getInputData();
@@ -66,34 +103,24 @@ export class JsonValidatorV1 implements INodeType {
 		// Loop through each input item and validate against the schema
 		const resultData: INodeExecutionData[] = [];
 
-		for (let i = 0; i < items.length; i++) {
-			const inputData = items[i].json[inputField];
+		for (const element of items) {
+			const inputData = element.json;
 
-			if (!inputData) {
-				throw new NodeOperationError(
-					this.getNode(),
-					`Input field "${inputField}" not found in item ${i}`,
-					{
-						itemIndex: i,
-					},
-				);
-			}
+			// Run validation
+			const isValid = validate(inputData);
 
-			//Run validation
-			const result = validate(inputData);
-
-			if (!result) {
+			if (!isValid) {
 				return this.prepareOutputData([
 					{
 						json: {
-							error: AJV.errorsText(validate.errors),
+							error: ajv.errorsText(validate.errors),
 						},
 					},
 				]);
 			}
 
 			// If valid, push the item to the output data
-			resultData.push(items[i]);
+			resultData.push(element);
 		}
 
 		// Return validated data
